pyproject.toml
[project]
name = "ade-bench"
version = "0.1.0"
description = "A benchmarking framework for data analyst tasks using dbt and SQL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "docker>=6.0",
    "libtmux>=0.23",
    "tqdm>=4.65",
    "click>=8.1",
    "rich>=13.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "boto3>=1.26",
    "psycopg2-binary>=2.9",
    "anthropic>=0.18",
    "openai>=1.0",
    "pandas>=2.0",
    "duckdb>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
wizard = "ade_bench.cli.wizard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true